Reorganize parametrization scripts
