+++ message_ix_models/model/material/data_util.py
@@ -1,8 +1,6 @@
-import os
 from functools import lru_cache
 from typing import TYPE_CHECKING, Literal, Mapping
 
-import ixmp
 import message_ix
 import numpy as np
 import pandas as pd
@@ -21,7 +19,6 @@ from message_ix_models.util import (
     broadcast,
     nodes_ex_world,
     package_data_path,
-    same_node,
 )
 
 if TYPE_CHECKING:
@@ -30,69 +27,7 @@ if TYPE_CHECKING:
     from message_ix_models import Context
 
 
-def load_GDP_COVID() -> pd.DataFrame:
-    """
-    Load COVID adjuste GDP projection
-
-    Returns
-    -------
-    pd.DataFrame
-    """
-    # Obtain 2015 and 2020 GDP values from NGFS baseline.
-    # These values are COVID corrected. (GDP MER)
-
-    mp = ixmp.Platform()
-    scen_NGFS = message_ix.Scenario(
-        mp, "MESSAGEix-GLOBIOM 1.1-M-R12-NGFS", "baseline", cache=True
-    )
-    gdp_covid_2015 = scen_NGFS.par("gdp_calibrate", filters={"year": 2015})
-
-    gdp_covid_2020 = scen_NGFS.par("gdp_calibrate", filters={"year": 2020})
-    gdp_covid_2020 = gdp_covid_2020.drop(["year", "unit"], axis=1)
-
-    # Obtain SSP2 GDP growth rates after 2020 (from ENGAGE baseline)
-
-    f_name = "iamc_db ENGAGE baseline GDP PPP.xlsx"
-
-    gdp_ssp2 = pd.read_excel(
-        package_data_path("material", "other", f_name), sheet_name="data_R12"
-    )
-    gdp_ssp2 = gdp_ssp2[gdp_ssp2["Scenario"] == "baseline"]
-    regions = "R12_" + gdp_ssp2["Region"]
-    gdp_ssp2 = gdp_ssp2.drop(
-        ["Model", "Scenario", "Unit", "Region", "Variable", "Notes"], axis=1
-    )
-    gdp_ssp2 = gdp_ssp2.loc[:, 2020:]
-    gdp_ssp2 = gdp_ssp2.divide(gdp_ssp2[2020], axis=0)
-    gdp_ssp2["node"] = regions
-    gdp_ssp2 = gdp_ssp2[gdp_ssp2["node"] != "R12_World"]
-
-    # Multiply 2020 COVID corrrected values with SSP2 growth rates
-
-    df_new = pd.DataFrame(columns=["node", "year", "value"])
-
-    for ind in gdp_ssp2.index:
-        df_temp = pd.DataFrame(columns=["node", "year", "value"])
-        region = gdp_ssp2.loc[ind, "node"]
-        mult_value = gdp_covid_2020.loc[
-            gdp_covid_2020["node"] == region, "value"
-        ].values[0]
-        temp = gdp_ssp2.loc[ind, 2020:2110] * mult_value
-        region_list = [region] * temp.size
-
-        df_temp["node"] = region_list
-        df_temp["year"] = temp.index
-        df_temp["value"] = temp.values
-
-        df_new = pd.concat([df_new, df_temp])
-
-    df_new["unit"] = "T$"
-    df_new = pd.concat([df_new, gdp_covid_2015])
-
-    return df_new
-
-
-def add_macro_COVID(
+def add_macro_materials(
     scen: message_ix.Scenario, filename: str, check_converge: bool = False
 ) -> message_ix.Scenario:
     """
@@ -135,611 +70,6 @@ def add_macro_COVID(
     scen = scen.add_macro(data, check_convergence=check_converge)
 
     return scen
-
-
-def modify_demand_and_hist_activity(scen: message_ix.Scenario) -> None:
-    """Take care of demand changes due to the introduction of material parents
-    Shed industrial energy demand properly.
-    Also need take care of remove dynamic constraints for certain energy carriers.
-    Adjust the historical activity of the related industry technologies
-    that provide output to different categories of industrial demand (e.g.
-    i_therm, i_spec, i_feed). The historical activity is reduced the same %
-    as the industrial demand is reduced.
-
-    Parameters
-    ----------
-    scen: message_ix.Scenario
-        scenario where industry demand should be reduced
-    """
-
-    # NOTE Temporarily modifying industrial energy demand
-    # From IEA database (dumped to an excel)
-
-    s_info = ScenarioInfo(scen)
-    fname = "MESSAGEix-Materials_final_energy_industry.xlsx"
-
-    if "R12_CHN" in s_info.N:
-        sheet_n = "R12"
-        region_type = "R12_"
-        region_name_CPA = "RCPA"
-        region_name_CHN = "CHN"
-    else:
-        sheet_n = "R11"
-        region_type = "R11_"
-        region_name_CPA = "CPA"
-        region_name_CHN = ""
-
-    df = pd.read_excel(
-        package_data_path("material", "other", fname), sheet_name=sheet_n, usecols="A:F"
-    )
-
-    # Filter the necessary variables
-    df = df[
-        (df["SECTOR"] == "feedstock (petrochemical industry)")
-        | (df["SECTOR"] == "feedstock (total)")
-        | (df["SECTOR"] == "industry (chemicals)")
-        | (df["SECTOR"] == "industry (iron and steel)")
-        | (df["SECTOR"] == "industry (non-ferrous metals)")
-        | (df["SECTOR"] == "industry (non-metallic minerals)")
-        | (df["SECTOR"] == "industry (total)")
-    ]
-    df = df[df["RYEAR"] == 2015]
-
-    # NOTE: Total cehmical industry energy: 27% thermal, 8% electricity, 65% feedstock
-    # SOURCE: IEA Sankey 2020: https://www.iea.org/sankey/#?c=World&s=Final%20consumption
-    # 67% of total chemicals energy is used for primary chemicals (ammonia,methnol,HVCs)
-    # SOURCE: https://www.iea.org/data-and-statistics/charts/primary-chemical-production-in-the-sustainable-development-scenario-2000-2030
-
-    # Retreive data for i_spec
-    # 67% of total chemcials electricity demand comes from primary chemicals (IEA)
-    # (Excludes petrochemicals as the share is negligable)
-    # Aluminum, cement and steel included.
-    # NOTE: Steel has high shares (previously it was not inlcuded in i_spec)
-
-    df_spec = df[
-        (df["FUEL"] == "electricity")
-        & (df["SECTOR"] != "industry (total)")
-        & (df["SECTOR"] != "feedstock (petrochemical industry)")
-        & (df["SECTOR"] != "feedstock (total)")
-    ]
-    df_spec_total = df[
-        (df["SECTOR"] == "industry (total)") & (df["FUEL"] == "electricity")
-    ]
-
-    df_spec_new = pd.DataFrame(
-        columns=["REGION", "SECTOR", "FUEL", "RYEAR", "UNIT_OUT", "RESULT"]
-    )
-    for r in df_spec["REGION"].unique():
-        df_spec_temp = df_spec.loc[df_spec["REGION"] == r]
-        df_spec_total_temp = df_spec_total.loc[df_spec_total["REGION"] == r]
-        df_spec_temp.loc[:, "i_spec"] = (
-            df_spec_temp.loc[:, "RESULT"]
-            / df_spec_total_temp.loc[:, "RESULT"].values[0]
-        )
-        df_spec_new = pd.concat([df_spec_temp, df_spec_new], ignore_index=True)
-
-    df_spec_new.drop(["FUEL", "RYEAR", "UNIT_OUT", "RESULT"], axis=1, inplace=True)
-    df_spec_new.loc[df_spec_new["SECTOR"] == "industry (chemicals)", "i_spec"] = (
-        df_spec_new.loc[df_spec_new["SECTOR"] == "industry (chemicals)", "i_spec"]
-        * 0.67
-    )
-
-    df_spec_new = df_spec_new.groupby(["REGION"]).sum().reset_index()
-
-    # Already set to zero: ammonia, methanol, HVCs cover most of the feedstock
-
-    df_feed = df[
-        (df["SECTOR"] == "feedstock (petrochemical industry)") & (df["FUEL"] == "total")
-    ]
-    # df_feed_total =
-    # df[(df["SECTOR"] == "feedstock (total)") & (df["FUEL"] == "total")]
-    df_feed_temp = pd.DataFrame(columns=["REGION", "i_feed"])
-    df_feed_new = pd.DataFrame(columns=["REGION", "i_feed"])
-
-    for r in df_feed["REGION"].unique():
-        i = 0
-        df_feed_temp.at[i, "REGION"] = r
-        df_feed_temp.at[i, "i_feed"] = 1
-        i = i + 1
-        df_feed_new = pd.concat([df_feed_temp, df_feed_new], ignore_index=True)
-
-    # Retreive data for i_therm
-    # 67% of chemical thermal energy chemicals comes from primary chemicals. (IEA)
-    # NOTE: Aluminum is excluded since refining process is not explicitly represented
-    # NOTE: CPA has a 3% share while it used to be 30% previosuly ??
-
-    df_therm = df[
-        (df["FUEL"] != "electricity")
-        & (df["FUEL"] != "total")
-        & (df["SECTOR"] != "industry (total)")
-        & (df["SECTOR"] != "feedstock (petrochemical industry)")
-        & (df["SECTOR"] != "feedstock (total)")
-        & (df["SECTOR"] != "industry (non-ferrous metals)")
-    ]
-    df_therm_total = df[
-        (df["SECTOR"] == "industry (total)")
-        & (df["FUEL"] != "total")
-        & (df["FUEL"] != "electricity")
-    ]
-    df_therm_total = (
-        df_therm_total.groupby(by="REGION").sum().drop(["RYEAR"], axis=1).reset_index()
-    )
-    df_therm = (
-        df_therm.groupby(by=["REGION", "SECTOR"])
-        .sum()
-        .drop(["RYEAR"], axis=1)
-        .reset_index()
-    )
-    df_therm_new = pd.DataFrame(
-        columns=["REGION", "SECTOR", "FUEL", "RYEAR", "UNIT_OUT", "RESULT"]
-    )
-
-    for r in df_therm["REGION"].unique():
-        df_therm_temp = df_therm.loc[df_therm["REGION"] == r]
-        df_therm_total_temp = df_therm_total.loc[df_therm_total["REGION"] == r]
-        df_therm_temp.loc[:, "i_therm"] = (
-            df_therm_temp.loc[:, "RESULT"]
-            / df_therm_total_temp.loc[:, "RESULT"].values[0]
-        )
-        df_therm_new = pd.concat([df_therm_temp, df_therm_new], ignore_index=True)
-        df_therm_new = df_therm_new.drop(["RESULT"], axis=1)
-
-    df_therm_new.drop(["FUEL", "RYEAR", "UNIT_OUT"], axis=1, inplace=True)
-    df_therm_new.loc[df_therm_new["SECTOR"] == "industry (chemicals)", "i_therm"] = (
-        df_therm_new.loc[df_therm_new["SECTOR"] == "industry (chemicals)", "i_therm"]
-        * 0.67
-    )
-
-    # Modify CPA based on https://www.iea.org/sankey/#?c=Japan&s=Final%20consumption.
-    # Since the value did not allign with the one in the IEA website.
-    index = (df_therm_new["SECTOR"] == "industry (iron and steel)") & (
-        (df_therm_new["REGION"] == region_name_CPA)
-        | (df_therm_new["REGION"] == region_name_CHN)
-    )
-
-    df_therm_new.loc[index, "i_therm"] = 0.2
-
-    df_therm_new = df_therm_new.groupby(["REGION"]).sum(numeric_only=True).reset_index()
-
-    # TODO: Useful technology efficiencies will also be included
-
-    # Add the modified demand and historical activity to the scenario
-
-    # Relted technologies that have outputs to useful industry level.
-    # Historical activity of theese will be adjusted
-    tec_therm = [
-        "biomass_i",
-        "coal_i",
-        "elec_i",
-        "eth_i",
-        "foil_i",
-        "gas_i",
-        "h2_i",
-        "heat_i",
-        "hp_el_i",
-        "hp_gas_i",
-        "loil_i",
-        "meth_i",
-        "solar_i",
-    ]
-    tec_fs = [
-        "coal_fs",
-        "ethanol_fs",
-        "foil_fs",
-        "gas_fs",
-        "loil_fs",
-        "methanol_fs",
-    ]
-    tec_sp = ["sp_coal_I", "sp_el_I", "sp_eth_I", "sp_liq_I", "sp_meth_I", "h2_fc_I"]
-
-    thermal_df_hist = scen.par("historical_activity", filters={"technology": tec_therm})
-    spec_df_hist = scen.par("historical_activity", filters={"technology": tec_sp})
-    feed_df_hist = scen.par("historical_activity", filters={"technology": tec_fs})
-    useful_thermal = scen.par("demand", filters={"commodity": "i_therm"})
-    useful_spec = scen.par("demand", filters={"commodity": "i_spec"})
-    useful_feed = scen.par("demand", filters={"commodity": "i_feed"})
-
-    for r in df_therm_new["REGION"]:
-        r_MESSAGE = region_type + r
-
-        useful_thermal.loc[useful_thermal["node"] == r_MESSAGE, "value"] = (
-            useful_thermal.loc[useful_thermal["node"] == r_MESSAGE, "value"]
-            * (1 - df_therm_new.loc[df_therm_new["REGION"] == r, "i_therm"].values[0])
-        )
-
-        thermal_df_hist.loc[thermal_df_hist["node_loc"] == r_MESSAGE, "value"] = (
-            thermal_df_hist.loc[thermal_df_hist["node_loc"] == r_MESSAGE, "value"]
-            * (1 - df_therm_new.loc[df_therm_new["REGION"] == r, "i_therm"].values[0])
-        )
-
-    for r in df_spec_new["REGION"]:
-        r_MESSAGE = region_type + r
-
-        useful_spec.loc[useful_spec["node"] == r_MESSAGE, "value"] = useful_spec.loc[
-            useful_spec["node"] == r_MESSAGE, "value"
-        ] * (1 - df_spec_new.loc[df_spec_new["REGION"] == r, "i_spec"].values[0])
-
-        spec_df_hist.loc[spec_df_hist["node_loc"] == r_MESSAGE, "value"] = (
-            spec_df_hist.loc[spec_df_hist["node_loc"] == r_MESSAGE, "value"]
-            * (1 - df_spec_new.loc[df_spec_new["REGION"] == r, "i_spec"].values[0])
-        )
-
-    for r in df_feed_new["REGION"]:
-        r_MESSAGE = region_type + r
-
-        useful_feed.loc[useful_feed["node"] == r_MESSAGE, "value"] = useful_feed.loc[
-            useful_feed["node"] == r_MESSAGE, "value"
-        ] * (1 - df_feed_new.loc[df_feed_new["REGION"] == r, "i_feed"].values[0])
-
-        feed_df_hist.loc[feed_df_hist["node_loc"] == r_MESSAGE, "value"] = (
-            feed_df_hist.loc[feed_df_hist["node_loc"] == r_MESSAGE, "value"]
-            * (1 - df_feed_new.loc[df_feed_new["REGION"] == r, "i_feed"].values[0])
-        )
-
-    scen.check_out()
-    scen.add_par("demand", useful_thermal)
-    scen.add_par("demand", useful_spec)
-    scen.add_par("demand", useful_feed)
-    scen.commit("Demand values adjusted")
-
-    scen.check_out()
-    scen.add_par("historical_activity", thermal_df_hist)
-    scen.add_par("historical_activity", spec_df_hist)
-    scen.add_par("historical_activity", feed_df_hist)
-    scen.commit(
-        comment="historical activity for useful level industry \
-    technologies adjusted"
-    )
-
-    # For aluminum there is no significant deduction required
-    # (refining process not included and thermal energy required from
-    # recycling is not a significant share.)
-    # For petro: based on 13.1 GJ/tonne of ethylene and the demand in the model
-
-    # df = scen.par('demand', filters={'commodity':'i_therm'})
-    # df.value = df.value * 0.38 #(30% steel, 25% cement, 7% petro)
-    #
-    # scen.check_out()
-    # scen.add_par('demand', df)
-    # scen.commit(comment = 'modify i_therm demand')
-
-    # Adjust the i_spec.
-    # Electricity usage seems negligable in the production of HVCs.
-    # Aluminum: based on IAI China data 20%.
-
-    # df = scen.par('demand', filters={'commodity':'i_spec'})
-    # df.value = df.value * 0.80  #(20% aluminum)
-    #
-    # scen.check_out()
-    # scen.add_par('demand', df)
-    # scen.commit(comment = 'modify i_spec demand')
-
-    # Adjust the i_feedstock.
-    # 45 GJ/tonne of ethylene or propylene or BTX
-    # 2020 demand of one of these: 35.7 Mt
-    # Makes up around 30% of total feedstock demand.
-
-    # df = scen.par('demand', filters={'commodity':'i_feed'})
-    # df.value = df.value * 0.7  #(30% HVCs)
-    #
-    # scen.check_out()
-    # scen.add_par('demand', df)
-    # scen.commit(comment = 'modify i_feed demand')
-
-    # NOTE Aggregate industrial coal demand need to adjust to
-    #      the sudden intro of steel setor in the first model year
-
-    t_i = ["coal_i", "elec_i", "gas_i", "heat_i", "loil_i", "solar_i"]
-
-    for t in t_i:
-        df = scen.par("growth_activity_lo", filters={"technology": t, "year_act": 2020})
-
-        scen.check_out()
-        scen.remove_par("growth_activity_lo", df)
-        scen.commit(comment="remove growth_lo constraints")
-
-    scen.check_out()
-    for substr in ["up", "lo"]:
-        df = scen.par(f"bound_activity_{substr}")
-        scen.remove_par(
-            f"bound_activity_{substr}",
-            df[(df["technology"].str.endswith("_fs")) & (df["year_act"] == 2020)],
-        )
-        scen.remove_par(
-            f"bound_activity_{substr}",
-            df[(df["technology"].str.endswith("_i")) & (df["year_act"] == 2020)],
-        )
-        scen.remove_par(
-            f"bound_activity_{substr}",
-            df[(df["technology"].str.endswith("_I")) & (df["year_act"] == 2020)],
-        )
-    scen.commit(comment="remove bounds")
-
-
-def modify_demand_and_hist_activity_debug(
-    scen: message_ix.Scenario,
-) -> dict[str, pd.DataFrame]:
-    """modularized "dry-run" version of modify_demand_and_hist_activity() for
-     debugging purposes
-
-    Parameters
-    ----------
-    scen: message_ix.Scenario
-        scenario to used to get i_therm and i_spec parametrization
-    Returns
-    ---------
-    dict[str, pd.DataFrame]
-        three keys named like MESSAGEix-GLOBIOM codes:
-            - i_therm
-            - i_spec
-            - historical_activity
-        values are DataFrames representing the reduced residual
-        industry demands when adding MESSAGEix-Materials to a
-        MESSAGEix-GLOBIOM scenario
-    """
-
-    s_info = ScenarioInfo(scen)
-    fname = "MESSAGEix-Materials_final_energy_industry.xlsx"
-
-    if "R12_CHN" in s_info.N:
-        sheet_n = "R12"
-        region_type = "R12_"
-        region_name_CPA = "RCPA"
-        region_name_CHN = "CHN"
-    else:
-        sheet_n = "R11"
-        region_type = "R11_"
-        region_name_CPA = "CPA"
-        region_name_CHN = ""
-
-    path = package_data_path("material", "other", fname)
-    df = pd.read_excel(path, sheet_name=sheet_n, usecols="A:F")
-
-    # Filter the necessary variables
-    df = df[
-        (df["SECTOR"] == "feedstock (petrochemical industry)")
-        | (df["SECTOR"] == "feedstock (total)")
-        | (df["SECTOR"] == "industry (chemicals)")
-        | (df["SECTOR"] == "industry (iron and steel)")
-        | (df["SECTOR"] == "industry (non-ferrous metals)")
-        | (df["SECTOR"] == "industry (non-metallic minerals)")
-        | (df["SECTOR"] == "industry (total)")
-    ]
-    df = df[df["RYEAR"] == 2015]
-
-    # NOTE: Total cehmical industry energy: 27% thermal, 8% electricity, 65% feedstock
-    # SOURCE: IEA Sankey 2020: https://www.iea.org/sankey/#?c=World&s=Final%20consumption
-    # 67% of total chemicals energy is used for primary chemicals (ammonia,methnol,HVCs)
-    # SOURCE: https://www.iea.org/data-and-statistics/charts/primary-chemical-production-in-the-sustainable-development-scenario-2000-2030
-
-    # Retreive data for i_spec
-    # 67% of total chemcials electricity demand comes from primary chemicals (IEA)
-    # (Excludes petrochemicals as the share is negligable)
-    # Aluminum, cement and steel included.
-    # NOTE: Steel has high shares (previously it was not inlcuded in i_spec)
-
-    df_spec = df[
-        (df["FUEL"] == "electricity")
-        & (df["SECTOR"] != "industry (total)")
-        & (df["SECTOR"] != "feedstock (petrochemical industry)")
-        & (df["SECTOR"] != "feedstock (total)")
-    ]
-    df_spec_total = df[
-        (df["SECTOR"] == "industry (total)") & (df["FUEL"] == "electricity")
-    ]
-
-    df_spec_new = pd.DataFrame(
-        columns=["REGION", "SECTOR", "FUEL", "RYEAR", "UNIT_OUT", "RESULT"]
-    )
-    for r in df_spec["REGION"].unique():
-        df_spec_temp = df_spec.loc[df_spec["REGION"] == r]
-        df_spec_total_temp = df_spec_total.loc[df_spec_total["REGION"] == r]
-        df_spec_temp.loc[:, "i_spec"] = (
-            df_spec_temp.loc[:, "RESULT"]
-            / df_spec_total_temp.loc[:, "RESULT"].values[0]
-        )
-        df_spec_new = pd.concat([df_spec_temp, df_spec_new], ignore_index=True)
-
-    df_spec_new.drop(["FUEL", "RYEAR", "UNIT_OUT", "RESULT"], axis=1, inplace=True)
-    df_spec_new.loc[df_spec_new["SECTOR"] == "industry (chemicals)", "i_spec"] = (
-        df_spec_new.loc[df_spec_new["SECTOR"] == "industry (chemicals)", "i_spec"]
-        * 0.67
-    )
-
-    df_spec_new = df_spec_new.groupby(["REGION"]).sum().reset_index()
-
-    # Already set to zero: ammonia, methanol, HVCs cover most of the feedstock
-
-    df_feed = df[
-        (df["SECTOR"] == "feedstock (petrochemical industry)") & (df["FUEL"] == "total")
-    ]
-    # df_feed_total =
-    # df[(df["SECTOR"] == "feedstock (total)") & (df["FUEL"] == "total")]
-    df_feed_temp = pd.DataFrame(columns=["REGION", "i_feed"])
-    df_feed_new = pd.DataFrame(columns=["REGION", "i_feed"])
-
-    for r in df_feed["REGION"].unique():
-        i = 0
-        df_feed_temp.at[i, "REGION"] = r
-        df_feed_temp.at[i, "i_feed"] = 1
-        i = i + 1
-        df_feed_new = pd.concat([df_feed_temp, df_feed_new], ignore_index=True)
-
-    # Retreive data for i_therm
-    # 67% of chemical thermal energy chemicals comes from primary chemicals. (IEA)
-    # NOTE: Aluminum is excluded since refining process is not explicitly represented
-    # NOTE: CPA has a 3% share while it used to be 30% previosuly ??
-
-    df_therm = df[
-        (df["FUEL"] != "electricity")
-        & (df["FUEL"] != "total")
-        & (df["SECTOR"] != "industry (total)")
-        & (df["SECTOR"] != "feedstock (petrochemical industry)")
-        & (df["SECTOR"] != "feedstock (total)")
-        & (df["SECTOR"] != "industry (non-ferrous metals)")
-    ]
-    df_therm_total = df[
-        (df["SECTOR"] == "industry (total)")
-        & (df["FUEL"] != "total")
-        & (df["FUEL"] != "electricity")
-    ]
-    df_therm_total = (
-        df_therm_total.groupby(by="REGION").sum().drop(["RYEAR"], axis=1).reset_index()
-    )
-    df_therm = (
-        df_therm.groupby(by=["REGION", "SECTOR"])
-        .sum()
-        .drop(["RYEAR"], axis=1)
-        .reset_index()
-    )
-    df_therm_new = pd.DataFrame(
-        columns=["REGION", "SECTOR", "FUEL", "RYEAR", "UNIT_OUT", "RESULT"]
-    )
-
-    for r in df_therm["REGION"].unique():
-        df_therm_temp = df_therm.loc[df_therm["REGION"] == r]
-        df_therm_total_temp = df_therm_total.loc[df_therm_total["REGION"] == r]
-        df_therm_temp.loc[:, "i_therm"] = (
-            df_therm_temp.loc[:, "RESULT"]
-            / df_therm_total_temp.loc[:, "RESULT"].values[0]
-        )
-        df_therm_new = pd.concat([df_therm_temp, df_therm_new], ignore_index=True)
-        df_therm_new = df_therm_new.drop(["RESULT"], axis=1)
-
-    df_therm_new.drop(["FUEL", "RYEAR", "UNIT_OUT"], axis=1, inplace=True)
-    df_therm_new.loc[df_therm_new["SECTOR"] == "industry (chemicals)", "i_therm"] = (
-        df_therm_new.loc[df_therm_new["SECTOR"] == "industry (chemicals)", "i_therm"]
-        * 0.67
-    )
-
-    # Modify CPA based on https://www.iea.org/sankey/#?c=Japan&s=Final%20consumption.
-    # Since the value did not allign with the one in the IEA website.
-    index = (df_therm_new["SECTOR"] == "industry (iron and steel)") & (
-        (df_therm_new["REGION"] == region_name_CPA)
-        | (df_therm_new["REGION"] == region_name_CHN)
-    )
-
-    df_therm_new.loc[index, "i_therm"] = 0.2
-
-    df_therm_new = df_therm_new.groupby(["REGION"]).sum(numeric_only=True).reset_index()
-
-    # TODO: Useful technology efficiencies will also be included
-
-    # Add the modified demand and historical activity to the scenario
-
-    # Relted technologies that have outputs to useful industry level.
-    # Historical activity of theese will be adjusted
-    tec_therm = [
-        "biomass_i",
-        "coal_i",
-        "elec_i",
-        "eth_i",
-        "foil_i",
-        "gas_i",
-        "h2_i",
-        "heat_i",
-        "hp_el_i",
-        "hp_gas_i",
-        "loil_i",
-        "meth_i",
-        "solar_i",
-    ]
-    tec_fs = [
-        "coal_fs",
-        "ethanol_fs",
-        "foil_fs",
-        "gas_fs",
-        "loil_fs",
-        "methanol_fs",
-    ]
-    tec_sp = ["sp_coal_I", "sp_el_I", "sp_eth_I", "sp_liq_I", "sp_meth_I", "h2_fc_I"]
-
-    thermal_df_hist = scen.par("historical_activity", filters={"technology": tec_therm})
-    spec_df_hist = scen.par("historical_activity", filters={"technology": tec_sp})
-    feed_df_hist = scen.par("historical_activity", filters={"technology": tec_fs})
-    useful_thermal = scen.par("demand", filters={"commodity": "i_therm"})
-    useful_spec = scen.par("demand", filters={"commodity": "i_spec"})
-    useful_feed = scen.par("demand", filters={"commodity": "i_feed"})
-
-    for r in df_therm_new["REGION"]:
-        r_MESSAGE = region_type + r
-
-        useful_thermal.loc[useful_thermal["node"] == r_MESSAGE, "value"] = (
-            useful_thermal.loc[useful_thermal["node"] == r_MESSAGE, "value"]
-            * (1 - df_therm_new.loc[df_therm_new["REGION"] == r, "i_therm"].values[0])
-        )
-
-        thermal_df_hist.loc[thermal_df_hist["node_loc"] == r_MESSAGE, "value"] = (
-            thermal_df_hist.loc[thermal_df_hist["node_loc"] == r_MESSAGE, "value"]
-            * (1 - df_therm_new.loc[df_therm_new["REGION"] == r, "i_therm"].values[0])
-        )
-
-    for r in df_spec_new["REGION"]:
-        r_MESSAGE = region_type + r
-
-        useful_spec.loc[useful_spec["node"] == r_MESSAGE, "value"] = useful_spec.loc[
-            useful_spec["node"] == r_MESSAGE, "value"
-        ] * (1 - df_spec_new.loc[df_spec_new["REGION"] == r, "i_spec"].values[0])
-
-        spec_df_hist.loc[spec_df_hist["node_loc"] == r_MESSAGE, "value"] = (
-            spec_df_hist.loc[spec_df_hist["node_loc"] == r_MESSAGE, "value"]
-            * (1 - df_spec_new.loc[df_spec_new["REGION"] == r, "i_spec"].values[0])
-        )
-
-    for r in df_feed_new["REGION"]:
-        r_MESSAGE = region_type + r
-
-        useful_feed.loc[useful_feed["node"] == r_MESSAGE, "value"] = useful_feed.loc[
-            useful_feed["node"] == r_MESSAGE, "value"
-        ] * (1 - df_feed_new.loc[df_feed_new["REGION"] == r, "i_feed"].values[0])
-
-        feed_df_hist.loc[feed_df_hist["node_loc"] == r_MESSAGE, "value"] = (
-            feed_df_hist.loc[feed_df_hist["node_loc"] == r_MESSAGE, "value"]
-            * (1 - df_feed_new.loc[df_feed_new["REGION"] == r, "i_feed"].values[0])
-        )
-
-    # For aluminum there is no significant deduction required
-    # (refining process not included and thermal energy required from
-    # recycling is not a significant share.)
-    # For petro: based on 13.1 GJ/tonne of ethylene and the demand in the model
-
-    # df = scen.par('demand', filters={'commodity':'i_therm'})
-    # df.value = df.value * 0.38 #(30% steel, 25% cement, 7% petro)
-    #
-    # scen.check_out()
-    # scen.add_par('demand', df)
-    # scen.commit(comment = 'modify i_therm demand')
-
-    # Adjust the i_spec.
-    # Electricity usage seems negligable in the production of HVCs.
-    # Aluminum: based on IAI China data 20%.
-
-    # df = scen.par('demand', filters={'commodity':'i_spec'})
-    # df.value = df.value * 0.80  #(20% aluminum)
-    #
-    # scen.check_out()
-    # scen.add_par('demand', df)
-    # scen.commit(comment = 'modify i_spec demand')
-
-    # Adjust the i_feedstock.
-    # 45 GJ/tonne of ethylene or propylene or BTX
-    # 2020 demand of one of these: 35.7 Mt
-    # Makes up around 30% of total feedstock demand.
-
-    # df = scen.par('demand', filters={'commodity':'i_feed'})
-    # df.value = df.value * 0.7  #(30% HVCs)
-    #
-    # scen.check_out()
-    # scen.add_par('demand', df)
-    # scen.commit(comment = 'modify i_feed demand')
-
-    # NOTE Aggregate industrial coal demand need to adjust to
-    #      the sudden intro of steel setor in the first model year
-    return {
-        "i_therm": useful_thermal,
-        "i_spec": useful_spec,
-        "historical_activity": pd.concat([thermal_df_hist, spec_df_hist, feed_df_hist]),
-    }
 
 
 def modify_baseyear_bounds(scen: message_ix.Scenario) -> None:
@@ -898,7 +228,7 @@ def calc_demand_shares(iea_db_df: pd.DataFrame, base_year: int) -> pd.DataFrame:
 
 
 def calc_resid_ind_demand(
-    scen: message_ix.Scenario, baseyear: int, iea_data_path: str
+    scen: message_ix.Scenario, baseyear: int, iea_data_path
 ) -> pd.DataFrame:
     comms = ["i_spec", "i_therm"]
     path = os.path.join(iea_data_path, "REV2022_allISO_IEA.parquet")
@@ -914,7 +244,7 @@ def calc_resid_ind_demand(
 
 
 def modify_industry_demand(
-    scen: message_ix.Scenario, baseyear: int, iea_data_path: str
+    scen: message_ix.Scenario, baseyear: int, iea_data_path
 ) -> None:
     df_demands_new = calc_resid_ind_demand(scen, baseyear, iea_data_path)
     scen.check_out()
@@ -1030,48 +360,224 @@ def read_iea_tec_map(tec_map_fname: str) -> pd.DataFrame:
     return MAP
 
 
-def get_hist_act_data(
-    map_fname: str, years: list or None = None, iea_data_path: str | None = None
+def read_sector_data(
+    scenario: message_ix.Scenario, sectname: str, ssp: str, filename: str
 ) -> pd.DataFrame:
     """
-    reads IEA DB, maps and aggregates variables to MESSAGE technologies
+    Read sector data for industry with sectname
 
     Parameters
     ----------
-    map_fname
-        name of MESSAGEix-technology-to-IEA-flow/product mapping file
-    years
-        specifies timesteps for whom historical activity should
-        be calculated and returned
+    scenario: message_ix.Scenario
+
+    sectname: sectname
+        name of industry sector
+
     Returns
     -------
     pd.DataFrame
 
     """
-    path = os.path.join(iea_data_path, "REV2022_allISO_IEA.parquet")
-    iea_enb_df = pd.read_parquet(path, engine="fastparquet")
-    if years:
-        iea_enb_df = iea_enb_df[iea_enb_df["TIME"].isin(years)]
+    # Read in technology-specific parameters from input xlsx
+    # Now used for steel and cement, which are in one file
 
-    # map IEA countries to MESSAGE region definition
-    iea_enb_df = map_iea_db_to_msg_regs(iea_enb_df)
+    import numpy as np
 
-    # read file for IEA product/flow - MESSAGE technologies map
-    MAP = read_iea_tec_map(map_fname)
+    s_info = ScenarioInfo(scenario)
 
-    # map IEA flows to MESSAGE technologies and aggregate
-    df_final = iea_enb_df.set_index(["PRODUCT", "FLOW"]).join(
-        MAP.set_index(["PRODUCT", "FLOW"])
+    if "R12_CHN" in s_info.N:
+        sheet_n = sectname + "_R12"
+    else:
+        sheet_n = sectname + "_R11"
+
+    # data_df = data_steel_china.append(data_cement_china, ignore_index=True)
+    data_df = pd.read_excel(
+        package_data_path("material", sectname, ssp, filename),
+        sheet_name=sheet_n,
     )
 
-    # multiply with efficiency and sector coverage ratios
+    # Clean the data
+    data_df = data_df[
+        [
+            "Region",
+            "Technology",
+            "Parameter",
+            "Level",
+            "Commodity",
+            "Mode",
+            "Species",
+            "Units",
+            "Value",
+        ]
+    ].replace(np.nan, "", regex=True)
 
-    df_final = (
-        df_final.drop_duplicates()
-        .groupby(["REGION", "technology", "TIME"])
-        .sum(numeric_only=True)
+    # Combine columns and remove ''
+    list_series = (
+        data_df[["Parameter", "Commodity", "Level", "Mode"]]
+        .apply(list, axis=1)
+        .apply(lambda x: list(filter(lambda a: a != "", x)))
     )
-    return df_final
+    list_ef = data_df[["Parameter", "Species", "Mode"]].apply(list, axis=1)
+
+    data_df["parameter"] = list_series.str.join("|")
+    data_df.loc[data_df["Parameter"] == "emission_factor", "parameter"] = (
+        list_ef.str.join("|")
+    )
+
+    data_df = data_df.drop(["Parameter", "Level", "Commodity", "Mode"], axis=1)
+    data_df = data_df.drop(data_df[data_df.Value == ""].index)
+
+    data_df.columns = data_df.columns.str.lower()
+
+    # Unit conversion
+
+    # At the moment this is done in the excel file, can be also done here
+    # To make sure we use the same units
+
+    return data_df
+
+
+def read_timeseries(
+    scenario: message_ix.Scenario, material: str, ssp: str or None, filename: str
+) -> pd.DataFrame:
+    """
+    Read "timeseries" type data from a sector specific xlsx input file
+    to DataFrame and format according to MESSAGEix standard
+
+    Parameters
+    ----------
+    ssp: str
+        if timeseries is available for different SSPs, the respective file is selected
+    scenario: message_ix.Scenario
+        scenario used to get structural information like
+        model regions and years
+    material: str
+        name of material folder where xlsx is located
+    filename:
+        name of xlsx file
+
+    Returns
+    -------
+    pd.DataFrame
+        DataFrame containing the timeseries data for MESSAGEix parameters
+    """
+    # Ensure config is loaded, get the context
+    s_info = ScenarioInfo(scenario)
+
+    # if context.scenario_info['scenario'] == 'NPi400':
+    #     sheet_name="timeseries_NPi400"
+    # else:
+    #     sheet_name = "timeseries"
+
+    if "R12_CHN" in s_info.N:
+        sheet_n = "timeseries_R12"
+    else:
+        sheet_n = "timeseries_R11"
+
+    material = f"{material}/{ssp}" if ssp else material
+    # Read the file
+
+    df = pd.read_excel(
+        package_data_path("material", material, filename), sheet_name=sheet_n
+    )
+
+    import numbers
+
+    # Take only existing years in the data
+    datayears = [x for x in list(df) if isinstance(x, numbers.Number)]
+
+    df = pd.melt(
+        df,
+        id_vars=[
+            "parameter",
+            "region",
+            "technology",
+            "mode",
+            "units",
+            "commodity",
+            "level",
+        ],
+        value_vars=datayears,
+        var_name="year",
+    )
+
+    df = df.drop(df[np.isnan(df.value)].index)
+    return df
+
+
+def read_rel(
+    scenario: message_ix.Scenario, material: str, ssp: str or None, filename: str
+) -> pd.DataFrame:
+    """
+    Read relation_* type parameter data for specific industry
+
+    Parameters
+    ----------
+    ssp: str
+        if relations are available for different SSPs, the respective file is selected
+    scenario:
+        scenario used to get structural information like
+    material: str
+        name of material folder where xlsx is located
+    filename:
+        name of xlsx file
+
+    Returns
+    -------
+    pd.DataFrame
+        DataFrame containing relation_* parameter data
+    """
+    # Ensure config is loaded, get the context
+
+    s_info = ScenarioInfo(scenario)
+
+    if "R12_CHN" in s_info.N:
+        sheet_n = "relations_R12"
+    else:
+        sheet_n = "relations_R11"
+    material = f"{material}/{ssp}" if ssp else material
+    # Read the file
+    data_rel = pd.read_excel(
+        package_data_path("material", material, filename),
+        sheet_name=sheet_n,
+    )
+
+    return data_rel
+
+
+def add_cement_ccs_co2_tr_relation(scen: message_ix.Scenario) -> None:
+    """Adds the relevant CCS technologies to the co2_trans_disp and bco2_trans_disp
+    relations
+
+    Parameters
+    ----------
+    scen: message_ix.Scenario
+        Scenario instance to add CCS emission factor parametrization to
+    """
+
+    # The relation coefficients for CO2_Emision and bco2_trans_disp and
+    # co2_trans_disp are both MtC. The emission factor for CCS add_ccs_technologies
+    # are specified in MtC as well.
+    co2_trans_relation = scen.par(
+        "emission_factor",
+        filters={
+            "technology": [
+                "clinker_dry_ccs_cement",
+                "clinker_wet_ccs_cement",
+            ],
+            "emission": "CO2",
+        },
+    )
+
+    co2_trans_relation.drop(["year_vtg", "emission", "unit"], axis=1, inplace=True)
+    co2_trans_relation["relation"] = "co2_trans_disp"
+    co2_trans_relation["node_rel"] = co2_trans_relation["node_loc"]
+    co2_trans_relation["year_rel"] = co2_trans_relation["year_act"]
+    co2_trans_relation["unit"] = "???"
+
+    scen.check_out()
+    scen.add_par("relation_activity", co2_trans_relation)
+    scen.commit("New CCS technologies added to the CO2 accounting relations.")
 
 
 def add_emission_accounting(scen: message_ix.Scenario) -> None:
@@ -1435,6 +941,7 @@ def add_emission_accounting(scen: message_ix.Scenario) -> None:
     # scen.commit("add methanol CO2_industry")
 
 
+def add_cement_bounds_2020(sc, temp: str = "high_temp"):
 def add_elec_lowerbound_2020(scen: message_ix.Scenario) -> None:
     # To avoid zero i_spec prices only for R12_CHN, add the below section.
     # read input parameters for relevant technology/commodity combinations for
@@ -1895,230 +1402,6 @@ def add_cement_bounds_2020(sc: message_ix.Scenario) -> None:
     sc.commit("added lower and upper bound for fuels for cement 2020.")
 
 
-def read_sector_data(
-    scenario: message_ix.Scenario, sectname: str, ssp: str
-) -> pd.DataFrame:
-    """
-    Read sector data for industry with sectname
-
-    Parameters
-    ----------
-    scenario: message_ix.Scenario
-
-    sectname: sectname
-        name of industry sector
-
-    Returns
-    -------
-    pd.DataFrame
-
-    """
-    # Read in technology-specific parameters from input xlsx
-    # Now used for steel and cement, which are in one file
-
-    import numpy as np
-
-    # Ensure config is loaded, get the context
-    context = read_config()
-
-    s_info = ScenarioInfo(scenario)
-
-    if "R12_CHN" in s_info.N:
-        sheet_n = sectname + "_R12"
-    else:
-        sheet_n = sectname + "_R11"
-
-    # data_df = data_steel_china.append(data_cement_china, ignore_index=True)
-    data_df = pd.read_excel(
-        package_data_path("material", "steel_cement", ssp, context.datafile),
-        sheet_name=sheet_n,
-    )
-
-    # Clean the data
-    data_df = data_df[
-        [
-            "Region",
-            "Technology",
-            "Parameter",
-            "Level",
-            "Commodity",
-            "Mode",
-            "Species",
-            "Units",
-            "Value",
-        ]
-    ].replace(np.nan, "", regex=True)
-
-    # Combine columns and remove ''
-    list_series = (
-        data_df[["Parameter", "Commodity", "Level", "Mode"]]
-        .apply(list, axis=1)
-        .apply(lambda x: list(filter(lambda a: a != "", x)))
-    )
-    list_ef = data_df[["Parameter", "Species", "Mode"]].apply(list, axis=1)
-
-    data_df["parameter"] = list_series.str.join("|")
-    data_df.loc[data_df["Parameter"] == "emission_factor", "parameter"] = (
-        list_ef.str.join("|")
-    )
-
-    data_df = data_df.drop(["Parameter", "Level", "Commodity", "Mode"], axis=1)
-    data_df = data_df.drop(data_df[data_df.Value == ""].index)
-
-    data_df.columns = data_df.columns.str.lower()
-
-    # Unit conversion
-
-    # At the moment this is done in the excel file, can be also done here
-    # To make sure we use the same units
-
-    return data_df
-
-
-def add_cement_ccs_co2_tr_relation(scen: message_ix.Scenario) -> None:
-    """Adds the relevant CCS technologies to the co2_trans_disp and bco2_trans_disp
-    relations
-
-    Parameters
-    ----------
-    scen: message_ix.Scenario
-        Scenario instance to add CCS emission factor parametrization to
-    """
-
-    # The relation coefficients for CO2_Emision and bco2_trans_disp and
-    # co2_trans_disp are both MtC. The emission factor for CCS add_ccs_technologies
-    # are specified in MtC as well.
-    co2_trans_relation = scen.par(
-        "emission_factor",
-        filters={
-            "technology": [
-                "clinker_dry_ccs_cement",
-                "clinker_wet_ccs_cement",
-            ],
-            "emission": "CO2",
-        },
-    )
-
-    co2_trans_relation.drop(["year_vtg", "emission", "unit"], axis=1, inplace=True)
-    co2_trans_relation["relation"] = "co2_trans_disp"
-    co2_trans_relation["node_rel"] = co2_trans_relation["node_loc"]
-    co2_trans_relation["year_rel"] = co2_trans_relation["year_act"]
-    co2_trans_relation["unit"] = "???"
-
-    scen.check_out()
-    scen.add_par("relation_activity", co2_trans_relation)
-    scen.commit("New CCS technologies added to the CO2 accounting relations.")
-
-
-# Read in time-dependent parameters
-def read_timeseries(
-    scenario: message_ix.Scenario, material: str, ssp: str or None, filename: str
-) -> pd.DataFrame:
-    """
-    Read "timeseries" type data from a sector specific xlsx input file
-    to DataFrame and format according to MESSAGEix standard
-
-    Parameters
-    ----------
-    ssp: str
-        if timeseries is available for different SSPs, the respective file is selected
-    scenario: message_ix.Scenario
-        scenario used to get structural information like
-        model regions and years
-    material: str
-        name of material folder where xlsx is located
-    filename:
-        name of xlsx file
-
-    Returns
-    -------
-    pd.DataFrame
-        DataFrame containing the timeseries data for MESSAGEix parameters
-    """
-    # Ensure config is loaded, get the context
-    s_info = ScenarioInfo(scenario)
-
-    # if context.scenario_info['scenario'] == 'NPi400':
-    #     sheet_name="timeseries_NPi400"
-    # else:
-    #     sheet_name = "timeseries"
-
-    if "R12_CHN" in s_info.N:
-        sheet_n = "timeseries_R12"
-    else:
-        sheet_n = "timeseries_R11"
-
-    material = f"{material}/{ssp}" if ssp else material
-    # Read the file
-
-    df = pd.read_excel(
-        package_data_path("material", material, filename), sheet_name=sheet_n
-    )
-
-    import numbers
-
-    # Take only existing years in the data
-    datayears = [x for x in list(df) if isinstance(x, numbers.Number)]
-
-    df = pd.melt(
-        df,
-        id_vars=[
-            "parameter",
-            "region",
-            "technology",
-            "mode",
-            "units",
-            "commodity",
-            "level",
-        ],
-        value_vars=datayears,
-        var_name="year",
-    )
-
-    df = df.drop(df[np.isnan(df.value)].index)
-    return df
-
-
-def read_rel(
-    scenario: message_ix.Scenario, material: str, ssp: str or None, filename: str
-) -> pd.DataFrame:
-    """
-    Read relation_* type parameter data for specific industry
-
-    Parameters
-    ----------
-    ssp: str
-        if relations are available for different SSPs, the respective file is selected
-    scenario:
-        scenario used to get structural information like
-    material: str
-        name of material folder where xlsx is located
-    filename:
-        name of xlsx file
-
-    Returns
-    -------
-    pd.DataFrame
-        DataFrame containing relation_* parameter data
-    """
-    # Ensure config is loaded, get the context
-
-    s_info = ScenarioInfo(scenario)
-
-    if "R12_CHN" in s_info.N:
-        sheet_n = "relations_R12"
-    else:
-        sheet_n = "relations_R11"
-    material = f"{material}/{ssp}" if ssp else material
-    # Read the file
-    data_rel = pd.read_excel(
-        package_data_path("material", material, filename),
-        sheet_name=sheet_n,
-    )
-
-    return data_rel
-
-
 def gen_te_projections(
     scen: message_ix.Scenario,
     ssp: Literal["all", "LED", "SSP1", "SSP2", "SSP3", "SSP4", "SSP5"] = "SSP2",
@@ -2264,125 +1547,6 @@ def calculate_ini_new_cap(
     return make_df("initial_new_capacity_up", **df_demand)
 
 
-def get_thermal_industry_emi_coefficients(scen: message_ix.Scenario) -> pd.DataFrame:
-    """
-    Pulls existing parametrization for non-CO2 emission
-    coefficients of given Scenario instance
-
-    Pulls MESSAGEix-GLOBIOM emission coefficients from "relation_activity"
-    and normalizes them to fuel inputs
-
-    Parameters
-    ----------
-    scen: message_ix.Scenario
-        Scenario instance to pull emission coefficients from
-    Returns
-    -------
-    pd.DataFrame
-    """
-    ind_th_tecs = ["biomass_i", "coal_i", "eth_i", "foil_i", "gas_i", "loil_i"]
-    relations = [
-        "BCA_Emission",
-        "CH4_Emission",
-        "CO_Emission",
-        "N2O_Emission",
-        "NH3_Emission",
-        "NOx_Emission",
-        "OCA_Emission",
-        "SO2_Emission",
-        "VOC_Emission",
-    ]
-    first_year = 2020
-    common_index = ["node_loc", "year_act", "technology"]
-    df_rel = scen.par(
-        "relation_activity", filters={"technology": ind_th_tecs, "relation": relations}
-    )
-    df_rel = df_rel[df_rel["year_act"].ge(first_year)]
-    df_rel = df_rel.set_index(common_index)
-
-    df_in = scen.par("input", filters={"technology": ind_th_tecs})
-    df_in = df_in[df_in["year_act"].ge(first_year)]
-    df_in = df_in.set_index(common_index)
-
-    df_joined = df_rel.join(
-        df_in[["value", "commodity"]], rsuffix="_in"
-    ).drop_duplicates()
-    df_joined["emi_factor"] = df_joined["value"] / df_joined["value_in"]
-    return df_joined
-
-
-def get_furnace_inputs(scen: message_ix.Scenario, first_year: int) -> pd.DataFrame:
-    """Return existing parametrization for input coefficients of given Scenario instance
-     and returns only for technologies with "furnace" in the name
-
-    Parameters
-    ----------
-    scen: message_ix.Scenario
-        Scenario instance to pull input parameter from
-    first_year: int
-        Earliest year for which furnace input parameter should be retrieved
-
-    Returns
-    -------
-    pd.DataFrame
-        a dataframe of furnace input paramter with index
-        ["node_loc", "year_act", "commodity", "technology"]
-    """
-    furn_tecs = "furnace"
-    df_furn = scen.par("input")
-    df_furn = df_furn[
-        (df_furn["technology"].str.contains(furn_tecs))
-        & (df_furn["year_act"].ge(first_year))
-    ]
-    df_furn = df_furn.set_index(["node_loc", "year_act", "commodity", "technology"])
-    return df_furn
-
-
-def calculate_furnace_non_co2_emi_coeff(
-    df_furn: pd.DataFrame, df_emi: pd.DataFrame
-) -> pd.DataFrame:
-    """
-    Joins input and emission coefficient DataFrames on
-    region, year and commodity to derive correct
-    non-CO2 emission factor parameter for industry furnaces
-
-    Parameters
-    ----------
-    df_furn: pd.DataFrame
-        DataFrame containing input parametrization of furnaces
-    df_emi: pd.DataFrame
-        DataFrame containing input normalized non-CO2 emission coefficients
-
-    Returns
-    -------
-    pd.DataFrame
-    """
-    df_final = (
-        pd.DataFrame(
-            df_emi.reset_index().set_index(["node_loc", "year_act", "commodity"])[
-                ["relation", "emi_factor"]
-            ]
-        )
-        .join(
-            df_furn["value"]
-            .reset_index()
-            .drop_duplicates()
-            .set_index(["node_loc", "year_act", "commodity", "technology"])
-        )
-        .reset_index()
-        .drop_duplicates()
-    )
-    df_final["value"] = df_final["value"] * df_final["emi_factor"]
-    df_final_new = (
-        make_df("relation_activity", **df_final)
-        .pipe(same_node)
-        .pipe(broadcast, mode=["high_temp", "low_temp"])
-    )
-    df_final_new["year_rel"] = df_final_new["year_act"]
-    df_final_new["unit"] = "???"
-    return df_final_new
-
-
 def calibrate_t_d_tecs(scenario: "Scenario"):
     # ------------------------------------------------------
     # Revise t_d historical_activity and 2020 activity bounds
@@ -2490,7 +1654,7 @@ def calibrate_t_d_tecs(scenario: "Scenario"):
         scenario.add_par("bound_activity_up", update_df)
 
 
-def maybe_add_water_tecs(scenario: "Scenario") -> None:
+def add_water_par_data(scenario: "Scenario") -> None:
     """Adds water supply technologies that are required for the Materials build
 
     Parameters
@@ -2498,20 +1662,14 @@ def maybe_add_water_tecs(scenario: "Scenario") -> None:
     scenario: .Scenario
         instance to check for water technologies and add if missing
     """
-    if "water_supply" not in list(scenario.set("level")):
-        scenario.check_out()
-        # add missing water tecs
-        scenario.add_set("technology", "extract__freshwater_supply")
-        scenario.add_set("level", "water_supply")
-        scenario.add_set("commodity", "freshwater_supply")
-
-        water_dict = pd.read_excel(
-            package_data_path("material", "other", "water_tec_pars.xlsx"),
-            sheet_name=None,
-        )
-        for par in water_dict.keys():
-            scenario.add_par(par, water_dict[par])
-        scenario.commit("add missing water tecs")
+    scenario.check_out()
+    water_dict = pd.read_excel(
+        package_data_path("material", "other", "water_tec_pars.xlsx"),
+        sheet_name=None,
+    )
+    for par in water_dict.keys():
+        scenario.add_par(par, water_dict[par])
+    scenario.commit("add missing water tecs")
 
 
 def calibrate_for_SSPs(scenario: "Scenario") -> None:
